Update base library to latest

+++ nativeui/gfx/win/native_theme.h
@@ -16,6 +16,7 @@
 
 #include <map>
 
+#include "base/macros.h"
 #include "nativeui/gfx/color.h"
 #include "nativeui/gfx/geometry/rect.h"
 #include "nativeui/gfx/geometry/size.h"
@@ -213,7 +214,7 @@ class NativeTheme {
   HMODULE theme_dll_;
 
   // A cache of open theme handles.
-  mutable HANDLE theme_handles_[Part::Count];
+  mutable HANDLE theme_handles_[static_cast<size_t>(Part::Count)];
 
   DISALLOW_COPY_AND_ASSIGN(NativeTheme);
 };
